requirements.txt
tesseract-ocr.exe open source engine
pip install pytesseract
pip install opencv-python
pytesseract.pytesseract.tesseract_cmd=r'<path for tesseract.exe>'
